refactor load_project into separate module
