+++ pypeman/graph.py
@@ -6,10 +6,34 @@ to be moved into this module
 """
 
 
+import importlib
+import sys
 import time
+import traceback
 
+from pypeman.conf import settings
 from pypeman import channels
 from pypeman.errors import PypemanError
+
+
+def load_project():
+    settings.init_settings()
+    project_module = settings.PROJECT_MODULE
+    try:
+        importlib.import_module(project_module)
+    except ImportError as exc:
+        msg = str(exc)
+        if 'No module' not in msg:
+            print("IMPORT ERROR %s" % project_module)
+            raise
+        if project_module not in msg:
+            print("IMPORT ERROR %s" % project_module)
+            raise
+        print("Missing '%s' module !" % project_module)
+        sys.exit(-1)
+    except Exception:
+        traceback.print_exc()
+        raise
 
 
 def wait_for_loop(tmax=5.0):
